Fix semi-random crashes
Somehow I was getting a situation where the block to a finally was getting deallocated before it would get used leading to an EXC_BAD_ACCESS.

I think it’s possibly because we refer to the block as `id` in various places, preventing ARC from realizing it is a block and that it needs a `-copy`.

The particular testcase I had would trigger it reliably, but it needs UIKit so I can’t add it to the tests.