Revert 5cd14e; Apply real fix
The real fix was in fact related to the CAAnimation itself retaining its delegate. The reason I at first did not realize this is because a *copy* of the CAAnimation is provided to the delegate callback. Both the original and copy have a delegate set. Trying to unset the delegate on the copy however crashes because it is apparently *immutable*.

If we then resolve our promise with that animation, then it strongly holds its delegate which strongly holds our PMKResolver which strongly holds the animation: a retain cycle: a memory leak.

Refs #196