Add `join` but mark it as unavailable
Because Little Bites of Cococa featured PromiseKit using this unavailable function we should provide any users that try to use it with a helpful pointer to when.

We cannot elegantly provide join for Swift, and frankly I’m not convinced it should have been added to our API at all: it breaks the first rule of promises: that errors propogate!