Show a log message for unexpected cancellation
The UIAlertView initializer encourages users to make alert views that have one button that is the cancel button. We don’t want to special case this, since it may be desired that such alerts cancel the chain, so lets show a log message since most of the time it is probably undesired for this to be a cancel alert/action-sheet.