Let `animated` and `completion` do something
Not sure why you're hard-coding values where the default parameters are more than sufficient. Why not let the caller decide whether he wants presentViewController() to animate, and/or do something when it's finished? Is this intentional? If so, the parameters should be removed entirely from the function definition, to prevent confusion.